✅ [TODO] Rename IPC patterns: agent:* -> agent.***\n\n- Update EventPattern strings in agent.controller to dot-style for consistency\n- mcp/bridge already use dot-style; no changes there
