--- apps/gui/src/main/agent/session/agent-session.module.ts
@@ -1,14 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AgentSessionService } from './agent-session.service';
-import { AgentSessionController } from './agent-session.controller';
-import { McpRegistryModule } from '../../common/mcp/mcp-registry.module';
-import { AgentCoreModule } from '../../common/agent/agent-core.module';
-import { AgentEventBridge } from '../events/agent-event-bridge';
-
-@Module({
-  imports: [McpRegistryModule, AgentCoreModule],
-  controllers: [AgentSessionController],
-  providers: [AgentSessionService, AgentEventBridge],
-  exports: [AgentSessionService],
-})
-export class AgentSessionModule {}
