+++ apps/gui/src/main/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { APP_PIPE } from '@nestjs/core';
 import { ValidationPipe } from '@nestjs/common';
-import { AgentSessionModule } from './agent/session/agent-session.module';
+import { AgentSessionModule } from './agent/agent.module';
 import { ElectronAppModule } from './electron/electron-app.module';
 import { PresetApiModule } from './preset/preset.module';
 import { McpUsageModule } from './common/mcp/mcp-usage.module';
